Avatar functionality is now on the Settings page. Now to drive the image saving using the SQLite local storage strategy.
