+++ src/pages/home/home.html
@@ -13,11 +13,11 @@
   <button ion-button color="secondary" (click)="loadQuiz()">Take a Quiz</button>
   <button ion-button color="primary" (click)="takePicture()">Take a Picture</button>
 
-  <ion-card>
+  <!-- <ion-card>
     <ion-card-content>
       <ion-card-title>Avatar</ion-card-title>
       <p>You look great.</p>
     </ion-card-content>
     <img id="myImage" src="https://avatars2.githubusercontent.com/u/15255538?v=3&u=0eb488826f13e1960b033d1fad50223ba529bebd&s=400" >
-  </ion-card>
+  </ion-card> -->
 </ion-content>
