+++ src/pages/home/home.ts
@@ -1,7 +1,6 @@
 import { Component } from '@angular/core';
 import { NavController } from 'ionic-angular';
 import { Quiz } from '../quiz/quiz';
-import { Camera, CameraOptions } from '@ionic-native/camera';
 
 @Component({
   selector: 'page-home',
@@ -9,9 +8,7 @@ import { Camera, CameraOptions } from '@ionic-native/camera';
 })
 export class HomePage {
 
-  public base64Image: string;
-
-  constructor(public navCtrl: NavController, private camera: Camera) {
+  constructor(public navCtrl: NavController) {
 
   }
 
@@ -21,28 +18,5 @@ export class HomePage {
 
   loadQuiz() {
     this.navCtrl.push(Quiz);
-  }
-
-  takePicture() {
-    const options: CameraOptions = {
-      quality: 100,
-      destinationType: this.camera.DestinationType.DATA_URL,
-      encodingType: this.camera.EncodingType.JPEG,
-      mediaType: this.camera.MediaType.PICTURE
-    }
-
-    this.camera.getPicture(options).then((imageData) => {
-      // imageData is either a base64 encoded string or a file URI
-      // If it's base64:
-      let base64Image = 'data:image/jpeg;base64,' + imageData;
-      console.log("image:");
-      console.log(base64Image);
-
-      // Populate the image with the data
-      (<HTMLImageElement>document.getElementById('myImage')).src = base64Image;
-    }, (err) => {
-      // Handle error
-      console.error(err);
-    });
   }
 }
